playground/python/ai-agent-trivia-game-python/requirements.txt
requests
python-dotenv
openai
pydantic